Move IconSize type into types folder.
