+++ lib/types/IconSize.ts
@@ -1,0 +1,1 @@
+export type IconSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
